+++ rasa_core/events/__init__.py
@@ -134,7 +134,17 @@ class UserUttered(Event):
         return UserUttered(None)
 
     def as_dict(self):
-        raise NotImplementedError()
+        return {"event": self.type_name,
+                "text": self.text,
+                "parse_data": self.parse_data}
+
+    @classmethod
+    def _from_parameters(cls, event_name, parameters, domain):
+        try:
+            return UserUttered.from_parse_data(parameters.get("text"),
+                                               parameters.get("parse_data"))
+        except KeyError as e:
+            raise ValueError("Failed to parse bot uttered event. {}".format(e))
 
     def as_story_string(self):
         if self.intent:
@@ -204,7 +214,7 @@ class BotUttered(Event):
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
         try:
-            return BotUttered(parameters["text"], parameters["data"])
+            return BotUttered(parameters.get("text"), parameters.get("data"))
         except KeyError as e:
             raise ValueError("Failed to parse bot uttered event. {}".format(e))
 
@@ -247,7 +257,7 @@ class TopicSet(Event):
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
         try:
-            return TopicSet(parameters["topic"])
+            return TopicSet(parameters.get("topic"))
         except KeyError as e:
             raise ValueError("Failed to parse set topic event. {}".format(e))
 
@@ -300,7 +310,7 @@ class SlotSet(Event):
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
         try:
-            return SlotSet(parameters["name"], parameters["value"])
+            return SlotSet(parameters.get("name"), parameters.get("value"))
         except KeyError as e:
             raise ValueError("Failed to parse set slot event. {}".format(e))
 
@@ -444,8 +454,8 @@ class ReminderScheduled(Event):
     def _from_story_string(cls, event_name, parameters, domain):
         logger.info("Reminders will be ignored during training, "
                     "which should be ok.")
-        return ReminderScheduled(parameters["action"],
-                                 parameters["date_time"],
+        return ReminderScheduled(parameters.get("action"),
+                                 parameters.get("date_time"),
                                  parameters.get("name", None),
                                  parameters.get("kill_on_user_msg", True))
 
@@ -562,6 +572,8 @@ class ActionExecuted(Event):
     It comprises an action and a list of events. operations will be appended
     to the latest ``Turn`` in the ``Tracker.turns``."""
 
+    type_name = "action"
+
     def __init__(self, action_name):
         self.action_name = action_name
         self.unpredictable = False
@@ -587,6 +599,14 @@ class ActionExecuted(Event):
             return ActionExecuted(event_name)
         else:
             return None
+
+    @classmethod
+    def _from_parameters(cls, event_name, parameters, domain):
+        return ActionExecuted(parameters.get("name"))
+
+    def as_dict(self):
+        return {"event": self.type_name,
+                "name": self.action_name}
 
     def apply_to(self, tracker):
         # type: (DialogueStateTracker) -> None