+++ rasa_core/server.py
@@ -9,11 +9,13 @@ import logging
 
 from builtins import str
 from klein import Klein
+from typing import Union, Text, Optional
 
 from rasa_core.agent import Agent
 from rasa_core.events import Event
+from rasa_core.interpreter import NaturalLanguageInterpreter
 from rasa_core.version import __version__
-from rasa_nlu.server import check_cors
+from rasa_nlu.server import check_cors, requires_auth
 
 logger = logging.getLogger(__name__)
 
@@ -41,6 +43,11 @@ def create_argument_parser():
             type=str,
             help="enable CORS for the passed origin. "
                  "Use * to whitelist all origins")
+    parser.add_argument(
+            '--auth_token',
+            type=str,
+            help="Enable token based authentication. Requests need to provide "
+                 "the token to be accepted.")
     parser.add_argument(
             '-o', '--log_file',
             type=str,
@@ -89,16 +96,23 @@ class RasaCoreServer(object):
                  loglevel="INFO",
                  log_file="rasa_core.log",
                  cors_origins=None,
-                 action_factory=None):
+                 action_factory=None,
+                 auth_token=None):
         logging.basicConfig(filename=log_file, level=loglevel)
         logging.captureWarnings(True)
 
-        self.config = {"cors_origins": cors_origins if cors_origins else []}
+        self.config = {"cors_origins": cors_origins if cors_origins else [],
+                       "token": auth_token}
         self.agent = self._create_agent(model_directory, interpreter,
                                         action_factory)
 
     @staticmethod
-    def _create_agent(model_directory, interpreter, action_factory=None):
+    def _create_agent(
+            model_directory,  # type: Text
+            interpreter,  # type: Union[Text, NaturalLanguageInterpreter]
+            action_factory=None #type: Optional[Text]
+    ):
+        # type: (...) -> Agent
         return Agent.load(model_directory, interpreter,
                           action_factory=action_factory)
 
@@ -110,6 +124,7 @@ class RasaCoreServer(object):
 
     @app.route("/conversations/<cid>/continue", methods=['POST', 'OPTIONS'])
     @check_cors
+    @requires_auth
     def continue_predicting(self, request, cid):
         request.setHeader('Content-Type', 'application/json')
         request_params = json.loads(
@@ -123,8 +138,54 @@ class RasaCoreServer(object):
                                                         events)
         return json.dumps(response)
 
+    @app.route("/conversations/<cid>/tracker/events", methods=['POST',
+                                                               'OPTIONS'])
+    @check_cors
+    def append_events(self, request, cid):
+        """Append a list of events to the state of a conversation"""
+        request.setHeader('Content-Type', 'application/json')
+        request_params = json.loads(
+                request.content.read().decode('utf-8', 'strict'))
+        events = convert_obj_2_tracker_events(request_params,
+                                              self.agent.domain)
+        tracker = self.agent.tracker_store.get_or_create_tracker(cid)
+        for e in events:
+            tracker.update(e)
+        self.agent.tracker_store.save(tracker)
+        return json.dumps(tracker.current_state())
+
+    @app.route("/conversations/<cid>/tracker", methods=['GET', 'OPTIONS'])
+    @check_cors
+    def retrieve_tracker(self, request, cid):
+        """Get a dump of a conversations tracker including its events."""
+
+        request.setHeader('Content-Type', 'application/json')
+        tracker = self.agent.tracker_store.get_or_create_tracker(cid)
+        return json.dumps(tracker.current_state(should_include_events=True))
+
+    @app.route("/conversations/<cid>/tracker", methods=['PUT', 'OPTIONS'])
+    @check_cors
+    def update_tracker(self, request, cid):
+        """Use a list of events to set a conversations tracker to a state."""
+
+        request.setHeader('Content-Type', 'application/json')
+        request_params = json.loads(
+                request.content.read().decode('utf-8', 'strict'))
+        events = convert_obj_2_tracker_events(request_params,
+                                              self.agent.domain)
+
+        tracker = self.agent.tracker_store.init_tracker(cid)
+        for e in events:
+            tracker.update(e)
+        self.agent.tracker_store.save(tracker)
+
+        # will override an existing tracker with the same id!
+        self.agent.tracker_store.save(tracker)
+        return json.dumps(tracker.current_state(should_include_events=True))
+
     @app.route("/conversations/<cid>/parse", methods=['GET', 'POST', 'OPTIONS'])
     @check_cors
+    @requires_auth
     def parse(self, request, cid):
         request.setHeader('Content-Type', 'application/json')
         if request.method.decode('utf-8', 'strict') == 'GET':
@@ -174,7 +235,8 @@ if __name__ == '__main__':
                           cmdline_args.nlu,
                           cmdline_args.loglevel,
                           cmdline_args.log_file,
-                          cmdline_args.cors)
+                          cmdline_args.cors,
+                          auth_token=cmdline_args.auth_token)
 
     logger.info("Started http server on port %s" % cmdline_args.port)
     rasa.app.run("0.0.0.0", cmdline_args.port)